Rename TypeScriptFile to TypeScriptFileA
