+++ packages/open-schema-type-script/src/example/datum-instance-type-script-configuration-definitions/testingPlatform/file/typeScriptFileA.ts
