+++ packages/open-schema-type-script/src/example/index.ts
@@ -37,7 +37,7 @@ import {
   buildTypeScriptFile,
   TypeScriptFile,
   TypeScriptFileTypeScriptConfiguration,
-} from './datum-instance-type-script-configuration-definitions/testingPlatform/file/typeScriptFile';
+} from './datum-instance-type-script-configuration-definitions/testingPlatform/file/typeScriptFileA';
 import { FileTypeScriptSemanticsIdentifier } from './datum-instance-type-script-configuration-definitions/testingPlatform/file/fileTypeScriptSemanticsIdentifier';
 import { FileExtensionSemanticsIdentifier } from './datum-instance-type-script-configuration-definitions/testingPlatform/file/fileExtensionSemanticsIdentifier';
